% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibrarySPoTApi.r
\name{LibrarySPoTApi_get_library_versions}
\alias{LibrarySPoTApi_get_library_versions}
\title{Retrieve a list of object versions by ID}
\usage{
LibrarySPoTApi_get_library_versions(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve a list of object versions by ID
}
